Refactor Sidebar; Fix ARIA issues
