+++ src/components/Footer/index.tsx
@@ -1,0 +1,2 @@
+export { Footer } from './Footer'
+// export type { FooterProps } from './Footer'
