+++ src/components/Header/Header.tsx
@@ -1,0 +1,64 @@
+import React from 'react'
+import { Grid, Toggle } from '~/components'
+
+export const Header = (): JSX.Element => (
+  <Grid container direction="row" justify="space-between" alignItems="center">
+    <svg width="220" height="33" viewBox="0 0 220 33" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <g clipPath="url(#clip0)">
+        <path d="M9.63669 0L0 33H4.25148L13.8882 0H9.63669Z" fill="#00EFE1" fillOpacity="0.95" />
+        <path d="M18.342 0L8.70532 33H12.9568L22.5935 0H18.342Z" fill="#00EFE1" fillOpacity="0.95" />
+        <path
+          d="M29.7686 32.695V10.4312H40.2393V13.7295H33.518V19.5016H38.5469V22.8H33.518V32.695H29.7686Z"
+          fill="white"
+        />
+        <path d="M42.4177 32.695V10.4312H46.159V29.3967H52.0706V32.695H42.4177Z" fill="white" />
+        <path
+          d="M58.3628 32.695H54.7349L58.7839 10.4312H63.4403L67.4893 32.695H63.8128L63.1488 27.7475H59.0268L58.3628 32.695ZM59.5775 24.4491H62.6062L61.1161 14.8427H61.0595L59.5775 24.4491Z"
+          fill="white"
+        />
+        <path d="M71.2629 13.7705H67.6511V10.4722H78.6726V13.7705H75.0366V32.736H71.2629V13.7705Z" fill="white" />
+        <path d="M81.3936 32.695V10.4312H85.143V32.695H81.3936Z" fill="white" />
+        <path
+          d="M92.4393 32.695H88.6899V10.4312H93.9375C98.1404 10.4312 100.294 12.2535 100.294 17.0279C100.294 20.6066 98.934 22.0578 97.6626 22.6433L100.837 32.7198H97.0228L94.3586 23.5009C93.7156 23.5683 93.0695 23.5986 92.4231 23.5916L92.4393 32.695ZM92.4393 20.4252H93.7755C95.8 20.4252 96.5532 19.6583 96.5532 17.0031C96.5532 14.3479 95.8 13.5811 93.7755 13.5811H92.4393V20.4252Z"
+          fill="white"
+        />
+        <path
+          d="M103.21 15.8322C103.21 12.6905 105.032 10.1343 108.959 10.1343C112.887 10.1343 114.717 12.6905 114.717 15.8322V27.3022C114.717 30.4439 112.895 33.0001 108.959 33.0001C105.024 33.0001 103.21 30.4439 103.21 27.3022V15.8322ZM106.967 27.2115C106.967 28.8607 107.453 29.6853 108.959 29.6853C110.466 29.6853 110.943 28.8607 110.943 27.2115V15.9229C110.943 14.2737 110.457 13.4491 108.943 13.4491C107.429 13.4491 106.951 14.2737 106.951 15.9229L106.967 27.2115Z"
+          fill="white"
+        />
+        <path
+          d="M118.345 32.695V10.4312H121.827L126.394 23.1628H126.459V10.4394H129.844V32.7033H126.605L121.827 18.8007H121.762V32.695H118.345Z"
+          fill="white"
+        />
+        <path
+          d="M145.822 15.923C145.668 14.0182 144.672 13.4492 143.821 13.4492C142.615 13.4492 141.951 14.2738 141.951 15.6426C141.951 19.4357 149.579 21.2498 149.579 27.2941C149.579 30.9552 147.15 32.992 143.643 32.992C140.137 32.992 138.169 30.1554 138.015 26.8323L141.586 26.3046C141.716 28.5639 142.615 29.6854 143.821 29.6854C144.101 29.7228 144.385 29.6977 144.653 29.6118C144.922 29.5259 145.169 29.3814 145.377 29.1882C145.586 28.9951 145.75 28.758 145.859 28.4935C145.968 28.229 146.019 27.9435 146.008 27.6569C146.008 23.2453 138.38 22.0744 138.38 15.7004C138.38 12.1546 140.493 10.1509 144.186 10.1509C147.247 10.1509 149.045 12.402 149.393 15.5437L145.822 15.923Z"
+          fill="white"
+        />
+        <path
+          d="M159.985 17.8689V16.1455C159.985 14.2655 159.135 13.4574 158.171 13.4574C156.657 13.4574 156.171 14.282 156.171 15.9311V27.2115C156.171 28.8607 156.657 29.6853 158.171 29.6853C159.532 29.6853 159.985 28.8607 159.985 27.4424V24.7377H163.743V27.3022C163.743 30.4439 161.921 33.0001 158.171 33.0001C154.236 33.0001 152.422 30.4439 152.422 27.3022V15.8322C152.422 12.6905 154.236 10.1343 158.171 10.1343C161.921 10.1343 163.743 12.9049 163.743 16.2362V17.8854L159.985 17.8689Z"
+          fill="white"
+        />
+        <path
+          d="M166.941 32.695V10.4312H170.699V19.5016H174.691V10.4312H178.449V32.695H174.691V22.8H170.699V32.695H166.941Z"
+          fill="white"
+        />
+        <path
+          d="M181.469 15.8322C181.469 12.6905 183.283 10.1343 187.219 10.1343C191.155 10.1343 192.969 12.6905 192.969 15.8322V27.3022C192.969 30.4439 191.155 33.0001 187.219 33.0001C183.283 33.0001 181.469 30.4439 181.469 27.3022V15.8322ZM185.227 27.2115C185.227 28.8607 185.705 29.6853 187.219 29.6853C188.733 29.6853 189.219 28.8607 189.219 27.2115V15.9229C189.219 14.2737 188.733 13.4491 187.219 13.4491C185.705 13.4491 185.227 14.2737 185.227 15.9229V27.2115Z"
+          fill="white"
+        />
+        <path
+          d="M195.973 15.8322C195.973 12.6905 197.787 10.1343 201.723 10.1343C205.658 10.1343 207.472 12.6905 207.472 15.8322V27.3022C207.472 30.4439 205.658 33.0001 201.723 33.0001C197.787 33.0001 195.973 30.4439 195.973 27.3022V15.8322ZM199.747 27.2115C199.747 28.8607 200.233 29.6853 201.747 29.6853C203.261 29.6853 203.747 28.8607 203.747 27.2115V15.9229C203.747 14.2737 203.261 13.4491 201.747 13.4491C200.233 13.4491 199.747 14.2737 199.747 15.9229V27.2115Z"
+          fill="white"
+        />
+        <path d="M210.339 32.695V10.4312H214.096V29.3967H220V32.695H210.339Z" fill="white" />
+      </g>
+      <defs>
+        <clipPath id="clip0">
+          <rect width="220" height="33" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+
+    <Toggle label="new student home experience" />
+  </Grid>
+)
