+++ src/components/Sidebar/index.tsx
@@ -1,0 +1,3 @@
+export { Sidebar } from './Sidebar'
+// export { SidebarNav } from './Sidebar/Nav'
+// export { SidebarNavItem } from './Sidebar/Nav/Item'
